content/SLC_script.js: turns out we do need to use loadImageCollection()
The problem with using `folder.loadFromUrl()` is that the imagesets in the
folder are not registered with the global list of imagesets, so
`setForgroundImageByName()` and friends won't be able to find the named
imagesets.

It ought to be possible to register the folder's imagesets with the global
list via the C# function `Wtml.LoadImagesets()`, but the Wtml class is private
and so isn't exposed to the JavaScript layer.

Fortunately, `ScriptInterface.loadImageCollection()` saves the folder object
that it creates in a private field, imageFolder, so we can hack a bit and pull
that out in the JS to gain access to the folder object. It would probably be
better to make the `Wtml` class public, but I don't want to jump to the v6
engine just yet.