Identify when description scrolls to bottom
When at bottom, scroll arrow disappears. However, it's present at times when it shouldn't be there at all.

Also, for whatever reason, the yellow border on the thumbnails isn't working anymore, and I have no idea why. I can't get $(this) to do anything on line 77 of the js file.